Revert "Merge branch 'Robin' into JohansBranch-(den-bästa)"
This reverts commit aa7ff562495732f4139409389dbf83d1ad1867d8, reversing
changes made to c44b69f0e6f8597a13c36bc154baa54f283ff935.

+++ Cheatscape/Main Menu.cs
@@ -42,7 +42,7 @@ namespace Cheatscape
         {
             if (animating == true)
             {
-                timer += gameTime.ElapsedGameTime.Milliseconds;
+                timer += (float)gameTime.ElapsedGameTime.Milliseconds;
 
                 if (timer >= framesPerSecond)
                 {
@@ -61,6 +61,7 @@ namespace Cheatscape
                 animating = false;
                 Global_Info.AccessCurrentGameState = Global_Info.GameState.LevelSelect;
             }
+
 
             if (CurrentFrame < 0)
             {
